Remove superfluous test class name parts
