+++ tests/Integration/SourceClient/DeleteTest.php
@@ -8,7 +8,7 @@ use SmartAssert\SourcesClient\Model\SourceInterface;
 use SmartAssert\SourcesClient\Tests\DataProvider\GetSourceDataProviderTrait;
 use SmartAssert\SourcesClient\Tests\Integration\AbstractIntegrationTestCase;
 
-class GetSourceTest extends AbstractIntegrationTestCase
+class DeleteTest extends AbstractIntegrationTestCase
 {
     use GetSourceDataProviderTrait;
 
@@ -24,11 +24,12 @@ class GetSourceTest extends AbstractIntegrationTestCase
      *
      * @param callable(): SourceInterface $creator
      */
-    public function testGetSourceSuccess(callable $creator): void
+    public function testDeleteSourceSuccess(callable $creator): void
     {
         $createdSource = $creator();
-        $retrievedSource = self::$sourceClient->get(self::$user1ApiToken->token, $createdSource->getId());
+        $retrievedSource = self::$sourceClient->delete(self::$user1ApiToken->token, $createdSource->getId());
 
-        self::assertEquals($createdSource, $retrievedSource);
+        self::assertSame($createdSource->getId(), $retrievedSource->getId());
+        self::assertNotNull($retrievedSource->getDeletedAt());
     }
 }
