+++ tests/Integration/SourceClient/GetTest.php
@@ -8,7 +8,7 @@ use SmartAssert\SourcesClient\Model\SourceInterface;
 use SmartAssert\SourcesClient\Tests\DataProvider\GetSourceDataProviderTrait;
 use SmartAssert\SourcesClient\Tests\Integration\AbstractIntegrationTestCase;
 
-class DeleteSourceTest extends AbstractIntegrationTestCase
+class GetTest extends AbstractIntegrationTestCase
 {
     use GetSourceDataProviderTrait;
 
@@ -24,12 +24,11 @@ class DeleteSourceTest extends AbstractIntegrationTestCase
      *
      * @param callable(): SourceInterface $creator
      */
-    public function testDeleteSourceSuccess(callable $creator): void
+    public function testGetSourceSuccess(callable $creator): void
     {
         $createdSource = $creator();
-        $retrievedSource = self::$sourceClient->delete(self::$user1ApiToken->token, $createdSource->getId());
+        $retrievedSource = self::$sourceClient->get(self::$user1ApiToken->token, $createdSource->getId());
 
-        self::assertSame($createdSource->getId(), $retrievedSource->getId());
-        self::assertNotNull($retrievedSource->getDeletedAt());
+        self::assertEquals($createdSource, $retrievedSource);
     }
 }
